% Generated by build script
\name{ons_dataset}
\alias{ons_dataset}
\title{Get ONS dataset details}
\usage{
ons_dataset()
}
\arguments{
\item{id}{Dataset ID (e.g. "wellbeing-quarterly", "cpih01")}
}
\value{tibble: edition, state, links_self, links_latest_version}
\description{Get ONS dataset details}
